+++ buffer/src/test/java/io/netty/buffer/api/tests/BufferSplitTest.java
@@ -46,25 +46,23 @@ public class BufferSplitTest extends BufferTestSupport {
     }
 
     private void readSplit(Fixture fixture, int capacity, int writeBytes, int readBytes, int offset) {
-        try (BufferAllocator allocator = fixture.createAllocator()) {
-            try (Buffer buf = allocator.allocate(capacity)) {
-                writeRandomBytes(buf, writeBytes);
-                assertEquals(writeBytes, buf.writerOffset());
+        try (BufferAllocator allocator = fixture.createAllocator(); Buffer buf = allocator.allocate(capacity)) {
+            writeRandomBytes(buf, writeBytes);
+            assertEquals(writeBytes, buf.writerOffset());
 
-                for (int i = 0; i < readBytes; i++) {
-                    buf.readByte();
-                }
-                assertEquals(readBytes, buf.readerOffset());
+            for (int i = 0; i < readBytes; i++) {
+                buf.readByte();
+            }
+            assertEquals(readBytes, buf.readerOffset());
 
-                try (Buffer split = buf.readSplit(offset)) {
-                    assertEquals(readBytes + offset, split.capacity());
-                    assertEquals(split.capacity(), split.writerOffset());
-                    assertEquals(readBytes, split.readerOffset());
+            try (Buffer split = buf.readSplit(offset)) {
+                assertEquals(readBytes + offset, split.capacity());
+                assertEquals(split.capacity(), split.writerOffset());
+                assertEquals(readBytes, split.readerOffset());
 
-                    assertEquals(capacity - split.capacity(), buf.capacity());
-                    assertEquals(writeBytes - split.capacity(), buf.writerOffset());
-                    assertEquals(0, buf.readerOffset());
-                }
+                assertEquals(capacity - split.capacity(), buf.capacity());
+                assertEquals(writeBytes - split.capacity(), buf.writerOffset());
+                assertEquals(0, buf.readerOffset());
             }
         }
     }
@@ -94,24 +92,22 @@ public class BufferSplitTest extends BufferTestSupport {
     }
 
     private void writeSplit(Fixture fixture, int capacity, int writeBytes, int readBytes, int offset) {
-        try (BufferAllocator allocator = fixture.createAllocator()) {
-            try (Buffer buf = allocator.allocate(capacity)) {
-                writeRandomBytes(buf, writeBytes);
-                assertEquals(writeBytes, buf.writerOffset());
-                for (int i = 0; i < readBytes; i++) {
-                    buf.readByte();
-                }
-                assertEquals(readBytes, buf.readerOffset());
+        try (BufferAllocator allocator = fixture.createAllocator(); Buffer buf = allocator.allocate(capacity)) {
+            writeRandomBytes(buf, writeBytes);
+            assertEquals(writeBytes, buf.writerOffset());
+            for (int i = 0; i < readBytes; i++) {
+                buf.readByte();
+            }
+            assertEquals(readBytes, buf.readerOffset());
 
-                try (Buffer split = buf.writeSplit(offset)) {
-                    assertEquals(writeBytes + offset, split.capacity());
-                    assertEquals(writeBytes, split.writerOffset());
-                    assertEquals(readBytes, split.readerOffset());
+            try (Buffer split = buf.writeSplit(offset)) {
+                assertEquals(writeBytes + offset, split.capacity());
+                assertEquals(writeBytes, split.writerOffset());
+                assertEquals(readBytes, split.readerOffset());
 
-                    assertEquals(capacity - split.capacity(), buf.capacity());
-                    assertEquals(0, buf.writerOffset());
-                    assertEquals(0, buf.readerOffset());
-                }
+                assertEquals(capacity - split.capacity(), buf.capacity());
+                assertEquals(0, buf.writerOffset());
+                assertEquals(0, buf.readerOffset());
             }
         }
     }
@@ -129,27 +125,25 @@ public class BufferSplitTest extends BufferTestSupport {
     }
 
     private static void splitPostFullOrRead(Fixture fixture, boolean read) {
-        try (BufferAllocator allocator = fixture.createAllocator()) {
-            final int capacity = 3;
-            try (Buffer buf = allocator.allocate(capacity)) {
-                writeRandomBytes(buf, capacity);
-                assertEquals(buf.capacity(), buf.writerOffset());
-                if (read) {
-                    for (int i = 0; i < capacity; i++) {
-                        buf.readByte();
-                    }
+        final int capacity = 3;
+        try (BufferAllocator allocator = fixture.createAllocator(); Buffer buf = allocator.allocate(capacity)) {
+            writeRandomBytes(buf, capacity);
+            assertEquals(buf.capacity(), buf.writerOffset());
+            if (read) {
+                for (int i = 0; i < capacity; i++) {
+                    buf.readByte();
                 }
-                assertEquals(read ? buf.capacity() : 0, buf.readerOffset());
+            }
+            assertEquals(read ? buf.capacity() : 0, buf.readerOffset());
 
-                try (Buffer split = buf.split()) {
-                    assertEquals(capacity, split.capacity());
-                    assertEquals(split.capacity(), split.writerOffset());
-                    assertEquals(read ? split.capacity() : 0, split.readerOffset());
+            try (Buffer split = buf.split()) {
+                assertEquals(capacity, split.capacity());
+                assertEquals(split.capacity(), split.writerOffset());
+                assertEquals(read ? split.capacity() : 0, split.readerOffset());
 
-                    assertEquals(0, buf.capacity());
-                    assertEquals(0, buf.writerOffset());
-                    assertEquals(0, buf.readerOffset());
-                }
+                assertEquals(0, buf.capacity());
+                assertEquals(0, buf.writerOffset());
+                assertEquals(0, buf.readerOffset());
             }
         }
     }
